requirements.txt
requests~=2.31.0
selenium~=4.12.0
fake-useragent=~1.3.0
python-dotenv~=1.0.0
python-anticaptcha~=1.0.0